read cache wip #1

+++ include/mind-net-api/mindnet/api/Trigger.hpp
@@ -41,8 +41,8 @@ namespace mindnet::api
             OperationResult& validation_result,
             OperationResult& action_result,
             model::ModelDefinition def,
-            int user_id,
-            int id,
+            i64 user_id,
+            i64 id,
             entity_fields& fields,
             entity_fields& old_fields,
             orm::QueryParams query_params
@@ -52,23 +52,23 @@ namespace mindnet::api
     int stack_depth,
     OperationResult& validation_result,
     const model::ModelDefinition& def,
-    int user_id,
-    int id,
+    i64 user_id,
+    i64 id,
     entity_fields& fields);
 
         virtual std::optional<std::pair<entity_fields, OperationResult>> run_instead_of_read(
             int stack_depth,
             OperationResult& validation_result,
             const model::ModelDefinition& def,
-            int user_id,
-            int id);
+            i64 user_id,
+            i64 id);
 
         virtual std::optional<OperationResult> run_instead_of_update(
             int stack_depth,
             OperationResult& validation_result,
             const model::ModelDefinition& def,
-            int user_id,
-            int id,
+            i64 user_id,
+            i64 id,
             entity_fields& fields,
             entity_fields& old_fields);
 
@@ -76,13 +76,13 @@ namespace mindnet::api
             int stack_depth,
             OperationResult& validation_result,
             const model::ModelDefinition& def,
-            int user_id,
-            int id);
+            i64 user_id,
+            i64 id);
         virtual std::optional<std::pair<std::vector<entity_fields>, OperationResult>> run_instead_of_list(
             int stack_depth,
             OperationResult& validation_result,
             const model::ModelDefinition& def,
-            int user_id,
+            i64 user_id,
             const orm::QueryParams& query_params);
 
         //
