+++ include/mind-net-api/mindnet/plugins/core/triggers/HistoryCommonTrigger.hpp
@@ -19,8 +19,8 @@ namespace mindnet::plugins::core::triggers
             api::OperationResult& validation_result,
             api::OperationResult& action_result,
             const model::ModelDefinition def,
-            int user_id,
-            int id,
+            i64 user_id,
+            i64 id,
             entity_fields& fields,
             entity_fields& old_fields,
             const orm::QueryParams query_params
